Decompile EntityRedEyeBust
