+++ Makefile
@@ -100,6 +100,8 @@ clean:
 format:
 	clang-format -i $$(find $(SRC_DIR)/ -type f -name "*.c")
 	clang-format -i $$(find $(INCLUDE_DIR)/ -type f -name "*.h")
+	# format saturn headers
+	clang-format -i $$(find $(SRC_DIR)/saturn -type f -name "*.h")
 	$(PYTHON) ./tools/symbols.py sort
 check:
 	sha1sum --check config/check.$(VERSION).sha
@@ -447,6 +449,7 @@ build_saturn_copy_files:
 	cp  ./src/saturn/t_bat.c $(SATURN_BUILD_DIR)
 	cp  ./src/saturn/zero.c $(SATURN_BUILD_DIR)
 	cp  ./src/saturn/stage_02.c $(SATURN_BUILD_DIR)
+	cp  ./src/saturn/stage_02.h $(SATURN_BUILD_DIR)
 	cp  ./src/saturn/sattypes.h $(SATURN_BUILD_DIR)
 	mkdir -p $(SATURN_BUILD_DIR)/asm/saturn/
 	mkdir -p $(SATURN_BUILD_DIR)/asm/saturn/
