+++ config/saturn/stage_02_user_syms.txt
@@ -15,3 +15,6 @@ _D_80180F10 =0x060F0E98;
 _SetStep =0x0607B218;
 _func_06079BB4 =0x06079BB4;
 _dat_060ed174 = 0x060F0E80;
+_entityRedEyeBustData2 = 0x060ecd70;
+_func_0600AFA8 = 0x0600afa8;
+_entityRedEyeBustData = 0x060ecf74;
