+++ src/saturn/stage_02.c
@@ -1,3 +1,4 @@
+#include "stage_02.h"
 #include "inc_asm.h"
 #include "sattypes.h"
 
@@ -5,7 +6,31 @@
 
 INCLUDE_ASM("asm/saturn/stage_02/data", d60DC000, d_060DC000);
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DC040, func_060DC040);
-INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DC460, func_060DC460);
+
+// SAT: func_060DC460
+// bust with red eyes that can have a candle on it
+void EntityRedEyeBust(Entity* self) {
+    struct Unk0600B344* result;
+    switch (self->step) {
+    case 0:
+        func_0607B264(self, 3);
+        result =
+            func_0600B344(entityRedEyeBustData.unk8, entityRedEyeBustData.unk10,
+                          entityRedEyeBustData.unk0, 1);
+        self->unk0 = result;
+        func_0600AFA8(result, entityRedEyeBustData2.unk28);
+        result->zPriority = 0x70;
+        result->unk14 = *(u32*)(&self->posX);
+        result->unk18 = *(u32*)(&self->posY);
+        self->step++;
+        break;
+    case 1:
+        break;
+    case 2:
+        break;
+    }
+}
+
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DC4EC, func_060DC4EC);
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DC510, func_060DC510);
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DC7B4, func_060DC7B4);
@@ -19,79 +44,19 @@ INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DD790, func_060DD790);
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DD8D8, func_060DD8D8);
 INCLUDE_ASM("asm/saturn/stage_02/f_nonmat", f60DDB80, func_060DDB80);
 
-extern u16 D_80180CC4[];
-extern const u8 D_80180EF0[];
-typedef enum {
-    E_NONE,
-    E_BREAKABLE,
-    E_EXPLOSION,
-    E_PRIZE_DROP,
-    E_NUMERIC_DAMAGE,
-    E_RED_DOOR,
-    E_INTENSE_EXPLOSION,
-    E_SOUL_STEAL_ORB,
-    E_ROOM_FOREGROUND,
-    E_STAGE_NAME_POPUP,
-    E_EQUIP_ITEM_DROP,
-    E_RELIC_ORB,
-    E_HEART_DROP,
-    E_ENEMY_BLOOD,
-    E_SAVE_GAME_POPUP,
-    E_DUMMY_0F,
-    E_DUMMY_10,
-
-    E_AXE_KNIGHT_AXE = 0x2A,
-    E_BONE_SCIMITAR_HEAD = 0x28,
-    E_FIRE = 0x38,
-    E_SLOGRA_SPEAR = 0x41,
-    E_SLOGRA_SPEAR_PROJECTILE = 0x42,
-    E_GAIBON = 0x43,
-    E_GAIBON_SMALL_FIREBALL = 0x45,
-    E_GAIBON_BIG_FIREBALL = 0x46,
-} EntityIDs;
-
-extern u16 D_80180F10[];
-extern const u8 D_80180EF8[];
-s32 AnimateEntity(Entity* entity, const u8 frames[], const u8 frames2[]);
-void InitializeEntity(u16 arg0[]);
-void func_801C29B0(s32 sfxId); // sfx
-void CreateEntityFromEntity(u16 objectId, Entity* source, Entity* entity);
-void SetStep(u8 step);
-
-void func_06079BB4(Entity*);
-void func_0607B264(Entity*, s32);
-
-struct Unk060ED26C {
-    s32 unk0;
-    u8 pad[4];
-    u16 unk8;
-    u16 unk10;
-};
-
-struct Unk0600B344 {
-    u8 pad[0xd];
-    s16 unke;
-};
-
-struct Unk0600B344* func_0600B344(s32, s32, s32, s32);
-extern struct Unk060ED26C D_060ED26C;
-extern u8 dat_060ed174[];
-
 // SAT: func_060DDE40
 // table with globe on it that can be broken
 void EntityTableWithGlobe(Entity* self) {
     struct Unk0600B344* result;
     switch (self->step) {
     case 0:
-
         func_0607B264(self, 5);
         self->step++;
-
         result = func_0600B344(D_060ED26C.unk8, D_060ED26C.unk10,
                                D_060ED26C.unk0, 7);
 
         self->unk0 = result;
-        result->unke = 0x6A;
+        result->zPriority = 0x6A;
         self->hitboxWidth = 10;
         self->hitboxHeight = 12;
         self->hitboxOffX = 0;
